requirements.txt
discord.py
yapf
